Add a new easy opponent level
Previous random placement of markers is now called the trivial
difficulty level.

The easy opponent looks at all of its options and chooses the one it
likes best. The evaluation function looks at possibilities of getting
many markers in a row and values longer combinations exponentially more.

It also does the same evaluation for the player and values their
possibilities negatively.

An obvious blind spot is not valuing gaps. For example XXX X is one move
away from win, but the evaluation function just sees possibilities of
getting 4 and 2 in a row for X. Neither of them can win, so there is no
need to worry in its perspective.